Builds. Still a few cleanup tweaks to do...

+++ src/api/routes/form-router.ts
@@ -2,7 +2,7 @@ import express, { Request, Response } from 'express';
 import { ReturnValidationErrors } from '../middleware';
 import { DB_CONFIG } from '../config';
 import knex from 'knex';
-import { UserService, FormService } from '../services';
+import { UserService, FormService, AuditService } from '../services';
 import { v4 as uuid } from 'uuid';
 import * as formHelper from '../utils/formHelper';
 import { auth } from 'express-openid-connect';
@@ -31,25 +31,36 @@ resetPgDateParsers();
 export const formRouter = express.Router();
 const userService = new UserService();
 const formService = new FormService();
+const auditService = new AuditService();
 
+// Get all forms for a user
 formRouter.get(
 	'/',
 	ReturnValidationErrors,
 	async function (req: Request, res: Response) {
 		try {
 			let user = await userService.getByEmail(req.user.email);
-			let auth = await db('forms').select('*').where('userId', '=', user.id);
+			let form = await db('forms').select('*').where('userId', '=', user.id);
 
-			for (let index = 0; index < auth.length; index++) {
-				auth[index].stops = await db('stops')
+			for (let index = 0; index < form.length; index++) {
+				form[index].stops = await db('stops')
 					.select('*')
-					.where('taid', '=', auth[index].id);
+					.where('taid', '=', form[index].id);
 				let departureDate = await db('stops')
 					.min('departureDate')
-					.where('taid', '=', auth[index].id);
-				auth[index].departureDate = departureDate[0].min;
+					.where('taid', '=', form[index].id);
+				let departureTime = await db('stops')
+					.select('departureTime')
+					.where('departureDate', '=', departureDate[0].min);
+
+				form[index].departureDate = departureDate[0].min
+					? departureDate[0].min
+					: 'Unknown';
+				form[index].departureTime = departureTime[0]
+					? departureTime[0].departureTime
+					: 'Unknown';
 			}
-			res.status(200).json(auth);
+			res.status(200).json(form);
 		} catch (error: any) {
 			console.log(error);
 			res.status(500).json('Internal Server Error');
@@ -122,6 +133,7 @@ formRouter.get(
 	}
 );
 
+//Get one of your own forms
 formRouter.get(
 	'/:formId',
 	ReturnValidationErrors,
@@ -129,17 +141,21 @@ formRouter.get(
 		try {
 			let user = await userService.getByEmail(req.user.email);
 
-			let auth = await db('forms')
+			let form = await db('forms')
 				.select('*')
 				.where('userId', '=', user.id)
 				.andWhere('formId', '=', req.params.formId)
 				.first();
 
-			if (auth) {
-				auth.stops = await db('stops').select('*').where('taid', '=', auth.id);
+			if (form) {
+				form.itinerary = await db('stops')
+					.select('*')
+					.where('taid', '=', form.id);
 
-				res.status(200).json(auth);
-			} else if (auth === undefined) {
+				console.log('form: ', form);
+
+				res.status(200).json(form);
+			} else if (form === undefined) {
 				res.status(200).json({ form: 'empty' });
 			} else {
 				res.status(404).json('Form not found');
@@ -156,15 +172,12 @@ formRouter.post(
 	'/:formId/save',
 	ReturnValidationErrors,
 	async function (req: Request, res: Response) {
-		console.log('Saving Form');
-
 		try {
 			await db.transaction(async (trx) => {
 				let user = await userService.getByEmail(req.user.email);
 
-				// let authInsert =
-				let stops = req.body.stops;
-				delete req.body.stops;
+				let itinerary = req.body.itinerary;
+				delete req.body.itinerary;
 
 				let authInsert = {
 					userId: user.id,
@@ -173,28 +186,42 @@ formRouter.post(
 					formId: req.params.formId,
 				};
 
-				let id = await db('forms')
-					.insert(authInsert, 'id')
-					.onConflict('formId')
-					.merge();
+				let form = await db('forms')
+					.select('*')
+					.andWhere('formId', '=', req.params.formId)
+					.first();
 
-				await db('stops')
-					.delete()
-					.where('taid', '=', id[0].id)
-					.transacting(trx);
+				if (!form || form.userId === user.id) {
+					let id = await db('forms')
+						.insert(authInsert, 'id')
+						.onConflict('formId')
+						.merge();
 
-				for (let index = 0; index < stops.length; index++) {
-					console.log('stop', stops[index]);
-					stops[index].travelTo = stops[index].travelTo.value;
-					stops[index].travelFrom = stops[index].travelFrom.value;
-					let stop = {
-						taid: id[0].id,
-						...stops[index],
-					};
-					await db('stops').insert(stop).transacting(trx);
+					await db('stops')
+						.delete()
+						.where('taid', '=', id[0].id)
+						.transacting(trx);
+
+					for (let index = 0; index < itinerary.length; index++) {
+						console.log('stop', itinerary[index]);
+						itinerary[index].travelTo = itinerary[index].locationId;
+						let stop = {
+							taid: id[0].id,
+							...itinerary[index],
+						};
+						await db('stops').insert(stop).transacting(trx);
+					}
+					auditService.insertAudit(
+						user.id,
+						id[0].id,
+						'Save',
+						'Successfully saved form'
+					);
+					res.status(200).json({ formId: req.params.formId });
+				} else {
+					res.status(401).json('Unauthorized');
 				}
 			});
-			res.status(200).json({ formId: req.params.formId });
 		} catch (error: any) {
 			console.log(error);
 			res.status(500).json('Insert failed');
@@ -263,6 +290,13 @@ formRouter.post(
 						};
 						await db('stops').insert(stop).transacting(trx);
 					}
+					auditService.insertAudit(
+						user.id,
+						id[0].id,
+						'Submit',
+						'Successfully submitted form'
+					);
+
 					res.status(200).json({ formId: req.params.formId });
 				} else {
 					res.status(500).json('Required fields in submission are blank');
@@ -305,6 +339,13 @@ formRouter.post(
 						.transacting(trx)
 						.returning('id');
 
+					auditService.insertAudit(
+						user.id,
+						id[0].id,
+						'Reassign',
+						'Successfully denied form'
+					);
+
 					res.status(200).json({ formId: req.body.formId });
 				} else {
 					res.status(500).json('Not authorized to deny this request');
@@ -341,6 +382,13 @@ formRouter.post(
 						.where('formId', '=', req.params.formId)
 						.transacting(trx)
 						.returning('id');
+
+					auditService.insertAudit(
+						user.id,
+						id[0].id,
+						'Reassign',
+						'Successfully approved form'
+					);
 
 					res.status(200).json({ formId: req.body.formId });
 				} else {
@@ -386,6 +434,12 @@ formRouter.post(
 						.transacting(trx)
 						.returning('id');
 
+					auditService.insertAudit(
+						user.id,
+						id[0].id,
+						'Reassign',
+						'Successfully reassigned form to ' + reassign
+					);
 					res.status(200).json({ formId: req.body.formId });
 				} else {
 					res.status(401).json('Must be supervisor to approve request');
@@ -425,6 +479,13 @@ formRouter.post(
 						.transacting(trx)
 						.returning('id');
 
+					auditService.insertAudit(
+						user.id,
+						id[0].id,
+						'Reassign',
+						'Successfully requested changes to form'
+					);
+
 					res.status(200).json({ formId: req.body.formId });
 				} else {
 					res.status(401).json('Must be supervisor to approve request');
@@ -444,17 +505,32 @@ formRouter.delete(
 	ReturnValidationErrors,
 	async function (req: Request, res: Response) {
 		try {
-			let result = await db('forms')
-				.update({
-					formstatus: 'deleted',
-				})
+			let user = await userService.getByEmail(req.user.email);
+
+			let id = await db('forms')
+				.select('id')
 				.where('formId', '=', req.params.formId)
-				.returning('formId');
-			if (result) {
-				res.status(200).json('Delete successful');
-				console.log('Delete successful', req.params.id);
+				.andWhere('email', '=', user.email)
+				.orWhere('supervisorEmail', '=', user.email);
+
+			if (id) {
+				let result = await db('forms')
+					.update({
+						formstatus: 'deleted',
+					})
+					.where('formId', '=', req.params.formId)
+					.returning('formId');
+
+				if (result) {
+					res.status(200).json('Delete successful');
+					console.log('Delete successful', req.params.id);
+
+					auditService.insertAudit(user.id, id[0].id, 'Delete', 'Deteled form');
+				} else {
+					res.status(500).json('Delete failed');
+				}
 			} else {
-				res.status(500).json('Delete failed');
+				res.status(401).json('Unauthorized');
 			}
 		} catch (error: any) {
 			res.status(500).json('Delete failed');
